updated torch cpp model loading
