--- examples/torch_script.cc
@@ -1,31 +1,0 @@
-#include <torch/script.h> // One-stop header.
-
-#include <iostream>
-#include <memory>
-
-int main(int argc, const char* argv[]) {
-  if (argc != 2) {
-    std::cerr << "usage: example-app <path-to-exported-script-module>\n";
-    return -1;
-  }
-
-  torch::jit::script::Module module;
-  try {
-    // Deserialize the ScriptModule from a file using torch::jit::load().
-    module = torch::jit::load(argv[1]);
-    // Create a vector of inputs.
-    std::vector<torch::jit::IValue> inputs;
-    inputs.push_back(torch::zeros({1,16}));
-    at::Tensor output = module.forward(inputs).toTensor();
-    auto max = torch::argmax(output);
-
-    auto maxNumber = *max.data<int64_t>();
-    std::cout << "Action:"<<maxNumber<<std::endl;
-  }
-  catch (const c10::Error& e) {
-    std::cerr << "error loading the model\n";
-    return -1;
-  }
-
-  std::cout << "ok\n";
-}
