+++ utils/docker/installers/requirements.txt
@@ -13,3 +13,4 @@ tensorflow>=2.2.0
 tensorboard>=2.2.2
 tf-agents>=0.5.0
 tensorflow-probability>=0.10.0
+torch>=1.5.0
